# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kenya Administrative Boundaries API(KenyaAdminAPI) - Documentation",
        "description": "The Kenya Administrative Boundaries API (KenAdminAPI) provides a comprehensive and easy-to-use interface for accessing detailed information about Kenya's administrative divisions, including Counties, Constituencies, and Wards. This API is designed to help developers integrate location-based data into their applications, supporting a wide range of use cases from geographical analysis to service delivery.",
        "contact": {
            "email": "sosmongare@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        },
        {
            "url": "https://127.0.0.1:8000",
            "description": "Local secure server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/api/counties": {
            "get": {
                "tags": [
                    "Counties"
                ],
                "summary": "Get list of counties",
                "operationId": "063497438427af6c02da6ea6f2dce272",
                "responses": {
                    "200": {
                        "description": "Counties retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/County"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Counties"
                ],
                "summary": "Create a new county",
                "operationId": "d6461e1ec40eedbd8f29323528557227",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "county_name"
                                ],
                                "properties": {
                                    "county_name": {
                                        "type": "string",
                                        "example": "Nairobi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "County created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/County"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/counties/{id}": {
            "get": {
                "tags": [
                    "Counties"
                ],
                "summary": "Get a county by ID",
                "operationId": "2c5dc583b4811058d596c5bdfa2e67e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "County retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/County"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Counties"
                ],
                "summary": "Update a county",
                "operationId": "7285c6fe82c79e5518ad13fcb5dbec34",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "county_name"
                                ],
                                "properties": {
                                    "county_name": {
                                        "type": "string",
                                        "example": "Nairobi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "County updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/County"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Counties"
                ],
                "summary": "Delete a county",
                "operationId": "3ce0c0a403e29c09d152e4dceb116a2e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "County deleted successfully"
                    }
                }
            }
        },
        "/api/counties/search/{query}": {
            "get": {
                "tags": [
                    "Counties"
                ],
                "summary": "Search for counties",
                "operationId": "8805d22e43fa945ef26254088926e0f4",
                "parameters": [
                    {
                        "name": "query",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/County"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No results found for the given query",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "County": {
                "title": "County",
                "description": "County model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of the county",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "county_name": {
                        "title": "County Name",
                        "description": "Name of the county",
                        "type": "string",
                        "example": "Nairobi"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "County"
                }
            },
...
        }
    },
...
}